move interface to io/runes
