+++ io/bytes/bytes.go
@@ -1,0 +1,5 @@
+package bytes
+
+import "io"
+
+type Reader = io.Reader
